extract module for managing double-sided tiles across titles
